Fjerner unødvendig kode
